Move common functions into inject_util, and implemented other code-review items
