+++ cli/cmd/root.go
@@ -76,13 +76,13 @@ func init() {
 	RootCmd.AddCommand(newCmdDashboard())
 	RootCmd.AddCommand(newCmdGet())
 	RootCmd.AddCommand(newCmdInject())
-	RootCmd.AddCommand(newCmdUninject())
 	RootCmd.AddCommand(newCmdInstall())
 	RootCmd.AddCommand(newCmdProfile())
 	RootCmd.AddCommand(newCmdRoutes())
 	RootCmd.AddCommand(newCmdStat())
 	RootCmd.AddCommand(newCmdTap())
 	RootCmd.AddCommand(newCmdTop())
+	RootCmd.AddCommand(newCmdUninject())
 	RootCmd.AddCommand(newCmdVersion())
 }
