+++ cli/cmd/testdata/inject_gettest_deployment.bad.golden
@@ -1,1 +1,1 @@
-Error injecting linkerd proxy: error converting YAML to JSON: yaml: line 14: did not find expected key
+Error transforming resources: error converting YAML to JSON: yaml: line 14: did not find expected key
